SmartHeart_Final/requirements.txt
flask==2.3.3
flask-session==0.5.0
werkzeug==2.3.7
scikit-learn==1.3.0
numpy==1.24.3
pandas==2.0.3
pillow==10.0.0
reportlab==4.0.4
pytesseract==0.3.10
opencv-python-headless==4.8.0.76
joblib==1.3.2
